Update pykalkan project version
The version in pyproject.toml was updated from 0.6.1 to 0.6.4. This update corresponds to significant changes in the structure and functionality of the project. Specifically, thread-safety has been added to the Adapter class and the Singleton design pattern has been implemented. This ensures multiple threads can interact safely with the Adapter class and only one instance of the Adapter class exists within the application, preventing accidental creation of new instances and maintaining a consistent state.