threading lock (#10)
* Implement Singleton pattern and add docstrings in adapter.py

Updated the Adapter class in adapter.py to use the singleton pattern, ensuring that only one instance of this class is created. This prevents multiple instances of this class from being created inadvertently, thus, maintaining a single state throughout the application's lifecycle.
Also, detailed docstrings were added to all public methods for better understanding of their functionalities. This increases maintainability and readability, providing clear explanations about what each method does, what are the input parameters and what they return.

* Add thread-safety to Adapter class and implement Singleton pattern

Added a lock to the Adapter class in adapter.py to ensure thread-safety and consistent state of the class instance. With threading.Lock(), multiple threads can safely instantiate and interact with the Adapter class without causing race conditions.
Also, updated Adapter class to use Singleton design pattern to ensure that only one instance exists within the application context. This prevents the accidental creation of multiple

* Update pykalkan project version

The version in pyproject.toml was updated from 0.6.1 to 0.6.4. This update corresponds to significant changes in the structure and functionality of the project. Specifically, thread-safety has been added to the Adapter class and the Singleton design pattern has been implemented. This ensures multiple threads can interact safely with the Adapter class and only one instance of the Adapter class exists within the application, preventing accidental creation of new instances and maintaining a consistent state.

* Update pykalkan to version 0.7.0

The project version in pyproject.toml was updated from 0.6.4 to 0.7.0. This update reflects major improvements in the project like the addition of thread-safety to the Adapter class and implementation of Singleton design pattern. These changes enhance consistency and safer interactions within the project.

* update lock var now inits in __new__func

* Update library name in main_test.py

* Refactor hardcoded data to variable in main_test.py