Testing (#11)
* Add `__exit__` method to clean up resources

Added a `__exit__` method to the `adapter.py` class to finalize the library and properly clean up resources when the instance is no longer in use. This is a preventive measure against potential memory leaks and is an enhancement towards more efficient resource handling.

* Remove SignatureFlag.KC_OUT_BASE64 from flags in lib_handle.py.

* Refactor main_test.py for better modularity

The previous structure of main_test.py was a bit haphazard with multiple levels of nested logic. It has been restructured to improve code modularity and readability. Code was split into smaller test functions, and constants were defined at the module level. A pytest fixture was also added to provide a common setup for all smaller test functions. Exception handling was adjusted to invoke pytest.fail directly.