thread safety and removing redundancies (#13)
* fix ValidateException

* version update

* threading lock (#10)

* Implement Singleton pattern and add docstrings in adapter.py

Updated the Adapter class in adapter.py to use the singleton pattern, ensuring that only one instance of this class is created. This prevents multiple instances of this class from being created inadvertently, thus, maintaining a single state throughout the application's lifecycle.
Also, detailed docstrings were added to all public methods for better understanding of their functionalities. This increases maintainability and readability, providing clear explanations about what each method does, what are the input parameters and what they return.

* Add thread-safety to Adapter class and implement Singleton pattern

Added a lock to the Adapter class in adapter.py to ensure thread-safety and consistent state of the class instance. With threading.Lock(), multiple threads can safely instantiate and interact with the Adapter class without causing race conditions.
Also, updated Adapter class to use Singleton design pattern to ensure that only one instance exists within the application context. This prevents the accidental creation of multiple

* Update pykalkan project version

The version in pyproject.toml was updated from 0.6.1 to 0.6.4. This update corresponds to significant changes in the structure and functionality of the project. Specifically, thread-safety has been added to the Adapter class and the Singleton design pattern has been implemented. This ensures multiple threads can interact safely with the Adapter class and only one instance of the Adapter class exists within the application, preventing accidental creation of new instances and maintaining a consistent state.

* Update pykalkan to version 0.7.0

The project version in pyproject.toml was updated from 0.6.4 to 0.7.0. This update reflects major improvements in the project like the addition of thread-safety to the Adapter class and implementation of Singleton design pattern. These changes enhance consistency and safer interactions within the project.

* update lock var now inits in __new__func

* Update library name in main_test.py

* Refactor hardcoded data to variable in main_test.py

* Testing (#11)

* Add `__exit__` method to clean up resources

Added a `__exit__` method to the `adapter.py` class to finalize the library and properly clean up resources when the instance is no longer in use. This is a preventive measure against potential memory leaks and is an enhancement towards more efficient resource handling.

* Remove SignatureFlag.KC_OUT_BASE64 from flags in lib_handle.py.

* Refactor main_test.py for better modularity

The previous structure of main_test.py was a bit haphazard with multiple levels of nested logic. It has been restructured to improve code modularity and readability. Code was split into smaller test functions, and constants were defined at the module level. A pytest fixture was also added to provide a common setup for all smaller test functions. Exception handling was adjusted to invoke pytest.fail directly.

* Adapter default arguments (#12)

* Refactor KalkanInterface methods in interface.py to improve usability

In interface.py, all methods of KalkanInterface were refactored. Additional parameters have been added to better specify the methods' functionalities. This will result in more robust error handling and improved reusability. New imports include CertCode, CertProp, and SignatureFlag enums from pykalkan.enums, and these are used as arguments in multiple methods. Additionally, the typing module is imported for type hinting purposes.

* Update methods' implementation in Adapter class for enhancement

The adapter.py in pykalkan has been reviewed and significant changes have been implemented. The KCClient import has been modified to LibHandle, and enums have been imported to facilitate input. KeyStore and data signature methods are updated to include flexibility in parameters through type hinting and default values. Additionally, CRL and OCSP validation have been refined to use ValidateType enum. The keys' store and data signing operations are revised to handle more complex use cases and improve convenience for users developing applications on top of pykalkan.

* Refactor functions in lib_handle.py to remove default parameters

The functions in lib_handle.py file have been refactored to remove default values from the parameters. This was done to ensure that essential parameters are not overlooked and to give users greater control over the function's behavior. It is intended to support varied usage and promote clear, self-explanatory code.

* Add OCSP URL parameter to x509_validate_certificate_ocsp

Included an additional parameter 'url' for the x509_validate_certificate_ocsp function to enable users to specify a different OCSP validation URL. Previously, the URL was hardcoded and lacked flexibility to check certificates against different OCSP services.

* Remove kalkan.py due to redundancies .

* Updated OCSP URL in main_test.py.

* Update version in pyproject.toml.

* Update README.md with detailed library usage

The README.md was updated to provide a detailed step-by-step instruction on how to use the Kalkan Crypt library. This includes connecting the library and loading the key store, signing data, and verifying the signature for validity and revocation using either OCSP or CRL. This ensures that users have clearer understanding and guidance when using the library.